Update optimization.py
Added each member's function with ability to choose their function. Added compute button for golden section method, which uses a member's golden section method to find a min/max. Added output text which shows the value of x where there is a min/max and the number of iterations it took to converge.